+++ changelog.md
@@ -11,6 +11,7 @@
 
 - Updated dependencies.
 - Also run GitHub Actions CI with Node.js v17.
+- Test the `processRequest` function with a [GraphQL multipart request](https://github.com/jaydenseric/graphql-multipart-request-spec) that has no files.
 
 ## 12.0.0